requirements.txt
streamlit==1.52.2
pandas==2.3.3
requests==2.32.5
plotly==5.24.1
matplotlib==3.10.0
wordcloud==1.9.4
emoji==2.14.0
nltk==3.9.1
python-dateutil==2.9.0.post0
